+++ src/neo4j/_async/bookmark_manager.py
@@ -19,7 +19,6 @@
 from __future__ import annotations
 
 import typing as t
-from collections import defaultdict
 
 from .._async_compat.concurrency import AsyncCooperativeLock
 from .._async_compat.util import AsyncUtil
@@ -29,9 +28,8 @@ from ..api import (
 )
 
 
-TBmSupplier = t.Callable[[t.Optional[str]],
-                         t.Union[Bookmarks, t.Awaitable[Bookmarks]]]
-TBmConsumer = t.Callable[[str, Bookmarks], t.Union[None, t.Awaitable[None]]]
+TBmSupplier = t.Callable[[], t.Union[Bookmarks, t.Awaitable[Bookmarks]]]
+TBmConsumer = t.Callable[[Bookmarks], t.Union[None, t.Awaitable[None]]]
 
 
 def _bookmarks_to_set(
@@ -45,62 +43,41 @@ def _bookmarks_to_set(
 class AsyncNeo4jBookmarkManager(AsyncBookmarkManager):
     def __init__(
         self,
-        initial_bookmarks: t.Optional[t.Mapping[str, t.Union[Bookmarks,
-                                                t.Iterable[str]]]] = None,
+        initial_bookmarks: t.Union[None, Bookmarks, t.Iterable[str]] = None,
         bookmarks_supplier: t.Optional[TBmSupplier] = None,
         bookmarks_consumer: t.Optional[TBmConsumer] = None
     ) -> None:
         super().__init__()
         self._bookmarks_supplier = bookmarks_supplier
         self._bookmarks_consumer = bookmarks_consumer
-        if initial_bookmarks is None:
-            initial_bookmarks = {}
-        self._bookmarks = defaultdict(
-            set, ((k, _bookmarks_to_set(v))
-                  for k, v in initial_bookmarks.items())
-        )
+        if not initial_bookmarks:
+            self._bookmarks = set()
+        else:
+            self._bookmarks = set(getattr(
+                initial_bookmarks, "raw_values",
+                t.cast(t.Iterable[str], initial_bookmarks)
+            ))
         self._lock = AsyncCooperativeLock()
 
     async def update_bookmarks(
-        self, database: str, previous_bookmarks: t.Collection[str],
+        self, previous_bookmarks: t.Collection[str],
         new_bookmarks: t.Collection[str]
     ) -> None:
         if not new_bookmarks:
             return
         with self._lock:
-            curr_bms = self._bookmarks[database]
-            curr_bms.difference_update(previous_bookmarks)
-            curr_bms.update(new_bookmarks)
+            self._bookmarks.difference_update(previous_bookmarks)
+            self._bookmarks.update(new_bookmarks)
             if self._bookmarks_consumer:
-                curr_bms_snapshot = Bookmarks.from_raw_values(curr_bms)
+                curr_bms_snapshot = Bookmarks.from_raw_values(self._bookmarks)
         if self._bookmarks_consumer:
-            await AsyncUtil.callback(
-                self._bookmarks_consumer, database, curr_bms_snapshot
-            )
+            await AsyncUtil.callback(self._bookmarks_consumer,
+                                     curr_bms_snapshot)
 
-    async def get_bookmarks(self, database: str) -> t.Set[str]:
+    async def get_bookmarks(self) -> t.Set[str]:
         with self._lock:
-            bms = set(self._bookmarks[database])
+            bms = set(self._bookmarks)
         if self._bookmarks_supplier:
-            extra_bms = await AsyncUtil.callback(
-                self._bookmarks_supplier, database
-            )
+            extra_bms = await AsyncUtil.callback(self._bookmarks_supplier)
             bms.update(extra_bms.raw_values)
         return bms
-
-    async def get_all_bookmarks(self) -> t.Set[str]:
-        bms: t.Set[str] = set()
-        with self._lock:
-            for database in self._bookmarks.keys():
-                bms.update(self._bookmarks[database])
-        if self._bookmarks_supplier:
-            extra_bms = await AsyncUtil.callback(
-                self._bookmarks_supplier, None
-            )
-            bms.update(extra_bms.raw_values)
-        return bms
-
-    async def forget(self, databases: t.Iterable[str]) -> None:
-        with self._lock:
-            for database in databases:
-                self._bookmarks.pop(database, None)